Move containers to fits.models.containers, add ReparametrizedModel, add SinMinMax
