+++ ionics_fits/models/sinusoid.py
@@ -1,7 +1,7 @@
 import numpy as np
 from typing import Dict, Tuple, TYPE_CHECKING
 
-from . import heuristics
+from . import heuristics, ReparametrizedModel
 from .. import common, Model, ModelParameter
 from ..utils import Array
 from . import utils
@@ -145,3 +145,47 @@ class Sinusoid(Model):
         )
 
         return derived_params, derived_uncertainties
+
+
+class SineMinMax(ReparametrizedModel):
+    """Sinusoid parametrised by minimum / maximum values instead of offset / amplitude.
+
+    This class is equivalent to :class Sinusoid: except that the `a` and `y0` parameters
+    are replaced with new `min` and `max` parameters defined by:
+      - `min = y0 - a`
+      - `max = y0 + a`
+
+    """
+
+    def __init__(self):
+        super().__init__(
+            model=Sinusoid,
+            new_params={
+                "min": ModelParameter(scale_func=common.scale_y),
+                "max": ModelParameter(scale_func=common.scale_y)
+            },
+            bound_params=["a", "y0"],
+        )
+
+    @staticmethod
+    def bound_param_values(param_values: dict[str, float]) -> dict[str, float]:
+        return {
+            "a": 0.5 * (param_values["max"] - param_values["min"]),
+            "y0": 0.5 * (param_values["max"] + param_values["min"]),
+        }
+
+    @staticmethod
+    def bound_param_uncertainties(
+        param_values: dict[str, float], param_uncertainties: dict[str, float]
+    ) -> dict[str, float]:
+        err = 0.5 * np.sqrt(
+            param_uncertainties["max"] ** 2 + param_uncertainties["min"] ** 2
+        )
+        return {"a": err, "y0": err}
+
+    @staticmethod
+    def new_param_values(model_param_values: dict[str, float]) -> dict[str, float]:
+        return {
+            "max": (model_param_values["y0"] + model_param_values["a"]),
+            "min": (model_param_values["y0"] - model_param_values["a"]),
+        }
